More rubust cache keys
When determining a cache key and stating that the ttl for the query should be now until for instance the end of day. We have a little tiny but in that the sha1 key will ALWAYS be different sense ttl value is stored along side the cache key. Use to the serialization of the object.

This pr fixes that.